+++ frontend/appflowy_flutter/lib/mobile/presentation/favorite/mobile_favorite_page.dart
@@ -82,16 +82,9 @@ class MobileFavoritePage extends StatelessWidget {
 
         // Folder
         Expanded(
-          child: Scrollbar(
-            child: SingleChildScrollView(
-              child: Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: MobileFavoritePageFolder(
-                  userProfile: userProfile,
-                  workspaceSetting: workspaceSetting,
-                ),
-              ),
-            ),
+          child: MobileFavoritePageFolder(
+            userProfile: userProfile,
+            workspaceSetting: workspaceSetting,
           ),
         ),
       ],
