+++ frontend/appflowy_flutter/lib/mobile/presentation/setting/support_setting_group.dart
@@ -4,6 +4,7 @@ import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 
 import 'widgets/widgets.dart';
 
@@ -14,32 +15,33 @@ class SupportSettingGroup extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MobileSettingGroup(
-      groupTitle: LocaleKeys.settings_mobile_support.tr(),
-      settingItemList: [
-        // 'Help Center'
-        MobileSettingItem(
-          name: LocaleKeys.settings_mobile_joinDiscord.tr(),
-          trailing: const Icon(
-            Icons.chevron_right,
+    return FutureBuilder(
+      future: PackageInfo.fromPlatform(),
+      builder: (context, snapshot) => MobileSettingGroup(
+        groupTitle: LocaleKeys.settings_mobile_support.tr(),
+        settingItemList: [
+          MobileSettingItem(
+            name: LocaleKeys.settings_mobile_joinDiscord.tr(),
+            trailing: const Icon(
+              Icons.chevron_right,
+            ),
+            onTap: () => safeLaunchUrl('https://discord.gg/JucBXeU2FE'),
           ),
-          onTap: () => safeLaunchUrl('https://discord.gg/JucBXeU2FE'),
-        ),
-        MobileSettingItem(
-          name: LocaleKeys.workspace_errorActions_reportIssue.tr(),
-          trailing: const Icon(
-            Icons.chevron_right,
+          MobileSettingItem(
+            name: LocaleKeys.workspace_errorActions_reportIssue.tr(),
+            trailing: const Icon(
+              Icons.chevron_right,
+            ),
+            onTap: () {
+              final String? version = snapshot.data?.version;
+              final String os = Platform.operatingSystem;
+              safeLaunchUrl(
+                'https://github.com/AppFlowy-IO/AppFlowy/issues/new?assignees=&labels=&projects=&template=bug_report.yaml&title=[Bug]%20Mobile:%20&version=$version&os=$os',
+              );
+            },
           ),
-          onTap: () {
-            // TODO(yijing): get app version before release
-            const String version = 'Beta';
-            final String os = Platform.operatingSystem;
-            safeLaunchUrl(
-              'https://github.com/AppFlowy-IO/AppFlowy/issues/new?assignees=&labels=&projects=&template=bug_report.yaml&title=[Bug]%20Mobile:%20&version=$version&os=$os',
-            );
-          },
-        ),
-      ],
+        ],
+      ),
     );
   }
 }
