+++ frontend/appflowy_flutter/lib/mobile/presentation/widgets/widgets.dart
@@ -1,3 +1,3 @@
 export 'show_flowy_mobile_confirm_dialog.dart';
 export 'show_flowy_mobile_bottom_sheet.dart';
-export 'flowy_mobile_error_state_container.dart';
+export 'flowy_mobile_state_container.dart';
